Initial import of tile portion of GeoPackage spec
Taken directly from the candidate standard released in http://www.opengeospatial.org/standards/requests/95 

Includes sections 10.1 to 10.5. May eventually need a bit more out of 10, on the metadata, but left out the raster portion, as we are focusing on tiles.

No formatting done on this commit, will attempt to clean up in the next commits.